master data doctype,custconttype and prodtype changes

+++ digilinq-account-service/src/main/java/com/mfino/digilinq/account/repository/DglMdDocTypeRepository.java
@@ -10,4 +10,4 @@ import com.mfino.digilinq.domain.DglMdDocType;
  */
 @SuppressWarnings("unused")
 @Repository
-public interface DglMdDocTypeRepository extends JpaRepository<DglMdDocType, Long> {}
+public interface DglMdDocTypeRepository extends JpaRepository<DglMdDocType, Integer> {}
